requirements.txt
python-gedcom>=1.0.0
sphinx>=7.0.0
sphinx-rtd-theme>=2.0.0
